test: one declaration for the scaffolding eleven suites had each written
Rule 2 applied to test fixtures, which is where a second copy is easiest to
justify and hardest to notice.

@dunx/openapi: generate-responses.test.ts carried a byte-for-byte copy of
generate-schemas.test.ts's three describes, so both suites ran the same
assertions twice. The copy is gone, and `operationOf` plus `info` - four copies
between them - are in document.fixture.ts.

@dunx/http: `withApp` was eleven copies of the same listen(0)/try/finally, and
a listen that is not closed on a failing assertion leaks a server into the rest
of the run. serving.fixture.ts owns the boot and the teardown; each suite keeps
its own `withApp` over it, because what actually differs is how the app is
built. The console capture was five copies with three different final filters.
relay.test.ts re-declared eight symbols the relay.fixture.ts beside it already
exported, and the two real-relay suites asserted cross-node delivery with the
same forty lines, now parameterised on the relay, the channel spelling and the
payload. The hand-written Standard Schema is in schema.fixture.ts.

@dunx/infra: the Redis reachability probe was four copies, including the four
client options that make it a probe rather than a connection.

core and infra each get a rejection.fixture.ts. It is deliberately not one copy
in @dunx/testing: architecture/packaging.md records that a published package's
tests importing a workspace package outside its runtime dependencies was tried
and reverted, because a package's build typechecks its own tests.

internal/bench: three dunx subjects shared a controller, which now lives in
servers/dunx-app.ts rather than servers/shared.ts. shared.ts is imported by all
twenty subjects, and an @dunx/http import in it would have put the framework in
the boot path of every one, corrupting the rss and startup numbers the harness
exists to produce. readme-tables.ts regenerates the README byte-identically.
int, micros and signed finish the consolidation format.ts was written for.

The relay work is verified against a live Redis and a live Postgres, not only
typechecked.
