+++ packages/http/src/server/input-reader.test.ts
@@ -2,14 +2,11 @@ import type { BunRequest } from 'bun';
 import { describe, expect, it } from 'bun:test';
 import { Module } from '@dunx/core';
 import { Controller, Post } from '../route/decorators.js';
-import type {
-  Input,
-  RouteSchemas,
-  StandardSchemaResult,
-  StandardSchemaV1,
-} from '../route/schema.js';
+import type { Input, RouteSchemas } from '../route/schema.js';
 import { ValidationError } from './errors.js';
+import { Note, field, schema } from './schema.fixture.js';
 import { HttpFactory, type HttpApp } from './factory.js';
+import { serving } from './serving.fixture.js';
 import { buildInputReader } from './input.js';
 import { HttpStatusCode } from './status.js';
 
@@ -19,27 +16,6 @@ import { HttpStatusCode } from './status.js';
  * doing it. A route that declares no schema must pay nothing for the feature
  * existing, and a synchronous validator must not cost a promise.
  */
-
-/** A Standard Schema by hand - the point being that no dependency is involved. */
-const schema = <T>(
-  validate: (
-    value: unknown,
-  ) => StandardSchemaResult<T> | Promise<StandardSchemaResult<T>>,
-): StandardSchemaV1<unknown, T> => ({
-  '~standard': { version: 1, vendor: 'test', validate },
-});
-
-const field = (value: unknown, key: string): unknown =>
-  typeof value === 'object' && value !== null
-    ? (value as Record<string, unknown>)[key]
-    : undefined;
-
-const Note = schema<{ text: string }>((value) => {
-  const text = field(value, 'text');
-  return typeof text === 'string'
-    ? { value: { text } }
-    : { issues: [{ message: 'text must be a string', path: ['text'] }] };
-});
 
 const Slow = schema<{ text: string }>(async (value) => {
   await Bun.sleep(1);
@@ -99,17 +75,9 @@ class NotesController {
 @Module({ controllers: [NotesController] })
 class AppModule {}
 
-const withApp = async (
+const withApp = (
   run: (app: HttpApp, url: string) => Promise<void>,
-): Promise<void> => {
-  const app = await HttpFactory.create(AppModule);
-  const url = await app.listen(0);
-  try {
-    await run(app, url);
-  } finally {
-    await app.shutdown();
-  }
-};
+): Promise<void> => serving(() => HttpFactory.create(AppModule), run);
 
 /**
  * Rejects everything, which is how "a response schema documents and never runs" is
