+++ packages/infra/src/cache/redis.test.ts
@@ -1,6 +1,7 @@
 import { AppFactory, type App } from '@dunx/core';
 import { afterAll, beforeAll, describe, expect, it } from 'bun:test';
 import { RedisConnection } from '../redis/connection.js';
+import { redisReachable } from '../reachable.fixture.js';
 import { RedisModule } from '../redis/module.js';
 import { defaultRedisUrl } from '../redis/options.js';
 import { FakeRedis } from './fake-redis.fixture.js';
@@ -52,25 +53,7 @@ describe('RedisCacheStore', () => {
 
 const url = defaultRedisUrl();
 
-const reachable = async (): Promise<boolean> => {
-  const client = new Bun.RedisClient(url, {
-    connectionTimeout: 500,
-    autoReconnect: false,
-    enableOfflineQueue: false,
-    maxRetries: 0,
-  });
-  try {
-    await client.connect();
-    await client.ping();
-    return true;
-  } catch {
-    return false;
-  } finally {
-    client.close();
-  }
-};
-
-const live = await reachable();
+const live = await redisReachable(url);
 if (!live) {
   console.log(`[dunx] cache integration tests skipped - ${url} unreachable`);
 }
