+++ packages/infra/src/queue/worker.test.ts
@@ -11,6 +11,7 @@ import {
 import type { Job } from 'bullmq';
 import { afterAll, beforeAll, describe, expect, it } from 'bun:test';
 import { defaultRedisUrl } from '../redis/options.js';
+import { redisReachable } from '../reachable.fixture.js';
 import { QueueConnection } from './connection.js';
 import { JobHandler } from './decorators.js';
 import { QueueError, QueueErrorCode } from './errors.js';
@@ -26,25 +27,8 @@ const url = defaultRedisUrl();
  * that needs one is conditional. Retries off, otherwise an unreachable host sits
  * in the offline queue for the default ten seconds.
  */
-const reachable = async (): Promise<boolean> => {
-  const client = new Bun.RedisClient(url, {
-    connectionTimeout: 500,
-    autoReconnect: false,
-    enableOfflineQueue: false,
-    maxRetries: 0,
-  });
-  try {
-    await client.connect();
-    await client.ping();
-    return true;
-  } catch {
-    return false;
-  } finally {
-    client.close();
-  }
-};
 
-const live = await reachable();
+const live = await redisReachable(url);
 if (!live) {
   console.log(`[dunx] queue integration tests skipped - ${url} unreachable`);
 }
