+++ packages/infra/src/redis/metrics.test.ts
@@ -1,6 +1,7 @@
 import { AppFactory, Module } from '@dunx/core';
 import { describe, expect, it } from 'bun:test';
 import { Redis } from './client.js';
+import { redisReachable } from '../reachable.fixture.js';
 import { RedisConnection } from './connection.js';
 import { RedisMetrics } from './metrics.js';
 import { redisMetrics, RedisModule } from './module.js';
@@ -19,25 +20,7 @@ const deadInit = {
 } as const;
 const offline = new RedisOptions(deadInit);
 
-const reachable = async (): Promise<boolean> => {
-  const client = new Bun.RedisClient(url, {
-    connectionTimeout: 500,
-    autoReconnect: false,
-    enableOfflineQueue: false,
-    maxRetries: 0,
-  });
-  try {
-    await client.connect();
-    await client.ping();
-    return true;
-  } catch {
-    return false;
-  } finally {
-    client.close();
-  }
-};
-
-const live = await reachable();
+const live = await redisReachable(url);
 
 describe('RedisMetrics', () => {
   it('keys one series per verb and totals across them', () => {
